Adding "apiType" as an option in selendroid capabilities.
Since some applications are written for Google API, they would fail to
install on an emulator with Android API, and vice versa. This implementation
will add "apiType" option to the selendroid capabilties to allow users
to select a specific emulator (Google API/Android API) with the desired
API type to run their tests.

Users can safely ignore this option if the application under test is not
sensitive to either API type to run.

Signed-off-by: Binh Vu <bvu@paypal.com>